Methods

Methods

This section describes systems methods implemented in IntervalMatrices.jl.

Methods

IntervalMatrices.left — Function.

left(A::IntervalMatrix)

The left part of this interval matrix, which corresponds to the left of each interval element in the matrix.

Input

A – interval matrix

source

IntervalMatrices.right — Function.

right(A::IntervalMatrix)

The right part of this interval matrix, which corresponds to the right of each interval element in the matrix.

Input

A – interval matrix

source

IntervalMatrices.quadratic_expansion — Function.

quadratic_expansion(A::IntervalMatrix, t, p)

Exactly compute the quadratic formula $At + 1/2A^2 * t^2$ using interval arithmetics.

Input

A – interval matrix

t – non-negative time value

Algorithm

See Lemma 1 in Reachability Analysis of Linear Systems with Uncertain Parameters and Inputs by M. Althoff, O. Stursberg, M. Buss.

source

IntervalMatrices.expm_overapproximation — Function.

expm_overapproximation(M::IntervalMatrix{T, <: AbstractInterval{T}}, t, p) where {T}

Overapproximation of the exponential of an interval matrix.

Input

A – interval matrix

t – non-negative time value

p – order of the appproximation

Algorithm

See Theorem 1 in Reachability Analysis of Linear Systems with Uncertain Parameters and Inputs by M. Althoff, O. Stursberg, M. Buss.

source

IntervalMatrices.expm_underapproximation — Function.

expm_underapproximation(M::IntervalMatrix{T, <: AbstractInterval{T}}, t, p) where {T}

Overapproximation of the exponential of an interval matrix.

Input

A – interval matrix

t – non-negative time value

p – order of the appproximation

Algorithm

See Theorem 2 in Reachability Analysis of Linear Systems with Uncertain Parameters and Inputs by M. Althoff, O. Stursberg, M. Buss.

source