Methods

Methods

This section describes systems methods implemented in IntervalMatrices.jl.

Methods

LinearAlgebra.opnorm — Function.

opnorm(A::IntervalMatrix, p::Real=Inf)

The matrix norm of an interval matrix.

Input

A – interval matrix

p – (optional, default: Inf): the class of p-norm

Notes

The matrix $p$-norm of an interval matrix $A$ is defined as

\[ ‖A‖_p := ‖\max(|\text{left}(A)|, |\text{right}(A)|)‖_p\]

where $\max$ and $|·|$ are taken elementwise.

source

IntervalMatrices.left — Function.

left(A::IntervalMatrix)

The left part of this interval matrix, which corresponds to the left of each interval element in the matrix.

Input

A – interval matrix

source

IntervalMatrices.right — Function.

right(A::IntervalMatrix)

The right part of this interval matrix, which corresponds to the right of each interval element in the matrix.

Input

A – interval matrix

source

IntervalMatrices.quadratic_expansion — Function.

quadratic_expansion(A::IntervalMatrix, t, p)

Exactly compute the quadratic formula $At + \frac{1}{2}A^2t^2$ using interval arithmetics.

Input

A – interval matrix

t – non-negative time value

Algorithm

See Lemma 1 in Reachability Analysis of Linear Systems with Uncertain Parameters and Inputs by M. Althoff, O. Stursberg, M. Buss.

source

IntervalMatrices.expm_overapproximation — Function.

expm_overapproximation(M::IntervalMatrix{T, <: AbstractInterval{T}}, t, p) where {T}

Overapproximation of the exponential of an interval matrix.

Input

A – interval matrix

t – non-negative time value

p – order of the appproximation

Algorithm

See Theorem 1 in Reachability Analysis of Linear Systems with Uncertain Parameters and Inputs by M. Althoff, O. Stursberg, M. Buss.

source

IntervalMatrices.expm_underapproximation — Function.

expm_underapproximation(M::IntervalMatrix{T, <: AbstractInterval{T}}, t, p) where {T}

Overapproximation of the exponential of an interval matrix.

Input

A – interval matrix

t – non-negative time value

p – order of the appproximation

Algorithm

See Theorem 2 in Reachability Analysis of Linear Systems with Uncertain Parameters and Inputs by M. Althoff, O. Stursberg, M. Buss.

source

Base.rand — Function.

rand(A::IntervalMatrix{T}; rng::AbstractRNG=GLOBAL_RNG) where {T}

Sample a concrete matrix $B$ from an interval matrix $A$ such that every entry in $B$ belongs to the corresponding interval in $A$.

Input

A – interval matrix

rng – (optional, default: GLOBAL_RNG) random-number generator

Output

A concrete matrix.

source

Base.split — Function.

split(A::IntervalMatrix{T}) where {T}

Split an interval matrix $A$ into two conventional matrices $C$ and $S$ such that $A = C + [-S, S]$.

Input

A – interval matrix

Output

A pair (C, S) such that the entries of C are the central points and the entries of S are the (nonnegative) radii of the intervals in A.

source

Missing docstring.

Missing docstring for ∈. Check Documenter's build log for details.